samples: tf-m: CBOR COSE decoder for non-secure side
The inferred data is encoded using CBOR, hashed and signed using COSE on the
secure side and transmitted to the non-secure side to parse to get back the
original inferred data payload, so added the decoder to parse and retrieve
the original inferred data payload.

Note:
This implementation is a kind of abstract on top of the NanoCBOR open-source
library APIs to decode the CBOR COSE payload.

Signed-off-by: Rajkumar Kanagaraj <rajkumar.kanagaraj@linaro.org>