test: arch: arm: Fix thread swap test clobbering
The thread swap test was making use of a register to point to
ztest_thread_callee_saved_regs_container, which was being clobbered
in the inline assembly code, causing a hard-fault in certain cases.

This commit assigns ztest_thread_callee_saved_regs_container to a
register (r2 for v6-m, r1 for v7-m/v8-m.mainline) earlier in the
inline assembly to ensure that the address isn't clobbered.

It also addresses an issue in the v6-m and v8-m.baseline inline
assembler where only 4 of the 8 callee-saved registers were being
set from ztest_thread_callee_saved_regs_container. With this commit,
the upper 4 registers are set with a combination of `ldr` + `mov`
(since `ldr` can not operate beyond r0-r7), before setting
the lower 4 callee-saved registers with `ldmia`.

Signed-off-by: Kevin Townsend <kevin.townsend@linaro.org>